# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend BlockID API",
    "description": "Read-only API for wallet trust scores and flags (data from database).",
    "version": "0.1.0"
  },
  "paths": {
    "/api/trust-score/{wallet}": {
      "get": {
        "tags": [
          "Trust Score",
          "trust-score"
        ],
        "summary": "Get Trust Score",
        "description": "Return trust score for a wallet. Reads from local database first (fast).\nReturns 404 if wallet has no cached score (sync worker or publish will populate).",
        "operationId": "get_trust_score_api_trust_score__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trust Score Api Trust Score  Wallet  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trust-score/list": {
      "post": {
        "tags": [
          "Trust Score",
          "trust-score"
        ],
        "summary": "List Trust Scores",
        "description": "Batch fetch trust scores from local database. One batch query; missing wallets\nreturn status=\"not_scored\". No RPC in hot path for fast response (< 50ms for 100 wallets).",
        "operationId": "list_trust_scores_api_trust_score_list_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustScoreListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Trust Scores Api Trust Score List Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/wallet/{address}": {
      "get": {
        "summary": "Get Wallet",
        "description": "Return the latest trust score and anomaly flags for a wallet.\n\nReads from database only. Returns 404 if the wallet has no trust score record.",
        "operationId": "get_wallet_wallet__address__get",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrustScoreListRequest": {
        "properties": {
          "wallets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Wallets",
            "description": "Wallet pubkeys to fetch trust scores for"
          }
        },
        "type": "object",
        "required": [
          "wallets"
        ],
        "title": "TrustScoreListRequest",
        "description": "POST /trust-score/list body: list of wallet pubkeys."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address",
            "description": "Wallet address (base58)"
          },
          "trust_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Trust Score",
            "description": "Latest trust score (0\u2013100)"
          },
          "computed_at": {
            "type": "integer",
            "title": "Computed At",
            "description": "Unix timestamp when score was computed"
          },
          "flags": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Flags",
            "description": "Anomaly flags from latest computation"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes",
            "description": "Trust score reason codes (e.g. NEW_WALLET, LOW_ACTIVITY)"
          }
        },
        "type": "object",
        "required": [
          "address",
          "trust_score",
          "computed_at"
        ],
        "title": "WalletResponse",
        "description": "GET /wallet/{address} response: latest trust score, flags, and reason codes from DB."
      },
...
    }
  }
}